Add files via upload
According to Osborne(Operations Reseach, 1959, pg145-172), due to Weber-Fechner Law, the price change of a stock is normal under logarithm. To check if such is true, the following code plots both price change and logarithmic price change. 
Where dictionary['close'] is stated, just insert a list of daily closing prices of a stock and adjust the number of elements in for i in range() statement.